test_bed/smolagent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smolagent"
version = "0.1.0"
description = "Multi-tool agent with CLI and web interface"
requires-python = ">=3.10"
dependencies = [
    "smolagents[litellm]",
    "duckduckgo-search",
    "requests",
    "beautifulsoup4",
    "slack-sdk",
    "python-dotenv",
    "prompt_toolkit",
    "gradio>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["smolagent"]